Fixed a bug that will cause the game flow to not be drawn when in exit state. Got rid of all of the warnings.

+++ SDL_Framework/Constants.cpp
@@ -16,7 +16,7 @@ void constants::setup::init_resolution_deps()
 void constants::setup::init_settings(std::vector<list_item> settings)
 {
 #pragma warning(disable:4244)
-	for (int i = 0; i < settings.size(); i++)
+	for (int i = 0; i < (int)settings.size(); i++)
 	{
 		if (settings.at(i).display_name == "resolution")
 		{
